+++ packages/react-components/react-headless-components-preview/library/src/components/InteractionTag/InteractionTagPrimary/InteractionTagPrimary.types.ts
@@ -1,0 +1,26 @@
+import type { InteractionTagPrimaryBaseState } from '@fluentui/react-tags';
+
+export type {
+  InteractionTagPrimarySlots,
+  InteractionTagPrimaryBaseProps as InteractionTagPrimaryProps,
+  InteractionTagPrimaryContextValues,
+} from '@fluentui/react-tags';
+
+export type InteractionTagPrimaryState = InteractionTagPrimaryBaseState & {
+  root: {
+    /**
+     * Data attribute set when the primary action is disabled.
+     */
+    'data-disabled'?: string;
+
+    /**
+     * Data attribute set when the interaction tag is selected.
+     */
+    'data-selected'?: string;
+
+    /**
+     * Data attribute set when the interaction tag has a secondary action.
+     */
+    'data-has-secondary-action'?: string;
+  };
+};