+++ packages/react-components/react-headless-components-preview/library/src/components/Tag/index.ts
@@ -1,0 +1,5 @@
+export { Tag } from './Tag';
+export { renderTag } from './renderTag';
+export { useTag } from './useTag';
+export { useTagContextValues } from './useTagContextValues';
+export type { TagSlots, TagProps, TagState, TagContextValues } from './Tag.types';